Stricter type checks (#169)
* update typescript config to be stricter

* initial approach to initializing entities in type-safe way

* fix entity initialization

* fix backend type issues

* fix entity initialization issue

* fix new emulator account handling

* update comment

* explicit stricter config in frontend

+++ backend/src/processes/managed-process.entity.ts
@@ -53,7 +53,7 @@ export class ManagedProcessEntity extends EventEmitter {
 
   async waitOnExit() {
     return new Promise<void>((resolve) => {
-      this.childProcess.once("exit", resolve);
+      this.childProcess?.once("exit", resolve);
     });
   }
 
@@ -65,7 +65,7 @@ export class ManagedProcessEntity extends EventEmitter {
     }
 
     this.logger.debug(
-      `Starting ${this.name} with command: ${command.name} ${command.args.join(
+      `Starting ${this.name} with command: ${command.name} ${command.args?.join(
         " "
       )}`
     );
@@ -86,7 +86,10 @@ export class ManagedProcessEntity extends EventEmitter {
     if (!this.isRunning()) {
       return;
     }
-    return new Promise<number>(async (resolve, reject) => {
+    return new Promise<number | null>(async (resolve, reject) => {
+      if (!this.childProcess) {
+        return;
+      }
       const isKilledSuccessfully = this.childProcess.kill("SIGINT");
       this.childProcess.once("error", (error) => {
         reject(error);
@@ -102,7 +105,7 @@ export class ManagedProcessEntity extends EventEmitter {
       const rejectionTimeoutSec = 6;
       setTimeout(() => {
         const timeoutError = new Error(
-          `Couldn't kill process ${this.id} (pid=${this.childProcess.pid}) within ${rejectionTimeoutSec}s timeout`
+          `Couldn't kill process ${this.id} (pid=${this.childProcess?.pid}) within ${rejectionTimeoutSec}s timeout`
         );
         reject(timeoutError);
       }, rejectionTimeoutSec * 1000);
@@ -118,7 +121,7 @@ export class ManagedProcessEntity extends EventEmitter {
     return {
       id: this.id,
       name: this.name,
-      command: { name, args },
+      command: { name, args: args ?? [] },
       state: this.state,
       output: this.output,
       updatedAt: this.updatedAt.toISOString(),
@@ -127,6 +130,9 @@ export class ManagedProcessEntity extends EventEmitter {
   }
 
   private attachEventListeners() {
+    if (!this.childProcess) {
+      return;
+    }
     this.childProcess.once("spawn", () => {
       this.logger.debug(`Process ${this.id} started`);
       this.setState(ManagedProcessState.MANAGED_PROCESS_STATE_RUNNING);
@@ -136,7 +142,7 @@ export class ManagedProcessEntity extends EventEmitter {
         `Process ${this.id} exited (code=${code}, signal=${signal})`
       );
       this.setState(
-        code > 0
+        code !== null && code > 0
           ? ManagedProcessState.MANAGED_PROCESS_STATE_ERROR
           : ManagedProcessState.MANAGED_PROCESS_STATE_NOT_RUNNING
       );
@@ -152,6 +158,9 @@ export class ManagedProcessEntity extends EventEmitter {
   }
 
   private detachEventListeners() {
+    if (!this.childProcess) {
+      return;
+    }
     // Make sure to remove all listeners to prevent memory leaks
     this.childProcess.stdout.removeAllListeners("data");
     this.childProcess.stderr.removeAllListeners("data");
